+++ src/components/ClientBar.svelte
@@ -1,18 +1,35 @@
+<script>
+    const companion = window.__COMPANION__;
+
+    let pinned = companion.isAlwaysOnTop();
+    let maximized = companion.isMaximized();
+
+    function pinToggle() {
+      companion.setAlwaysOnTop();
+      pinned = companion.isAlwaysOnTop();
+    }
+
+    function maximizeToggle() {
+      companion.maximizeToggle();
+      maximized = companion.isMaximized();
+    }
+</script>
+
 <div class="client-bar">
   <nav class="client-m drag">
     <div class="client-m__btns--left">
-      <span class="client-m__btn click pin-btn">
-        <i class="fa fa-unlock js-pin"></i>
+      <span class="client-m__btn click" on:click={pinToggle}>
+        <i class="fa {pinned ? 'fa-lock' : 'fa-unlock'}"></i>
       </span>
     </div>
     <div class="client-m__btns--right">
-      <span class="client-m__btn click minimize-btn">
+      <span class="client-m__btn click" on:click={companion.minimize}>
         <i class="fa fa-window-minimize"></i>
       </span>
-      <span class="client-m__btn click maximize-btn">
-        <i class="fa fa-window-maximize"></i>
+      <span class="client-m__btn click" on:click={maximizeToggle}>
+        <i class="fa {maximized ? 'fa-window-restore' : 'fa-window-maximize'}"></i>
       </span>
-      <span class="client-m__btn click close-btn">
+      <span class="client-m__btn click" on:click={companion.close}>
         <i class="fa fa-times"></i>
       </span>
     </div>