+++ src/components/WebView.svelte
@@ -1,8 +1,10 @@
 <script>
   import { url, webview } from './store.js';
+
+  export let expand;
 </script>
 
-<div class="webview">
+<div class={ expand? 'webview webview--full' : 'webview' }>
   <webview 
     bind:this={$webview}
     src={$url}